requirements.txt
duckdb
ibis-framework
ibis-substrait
protobuf
--extra-index-url https://pypi.fury.io/arrow-nightlies --prefer-binary --pre
pyarrow
pytest
substrait-validator